route CRUD and gate way

+++ client/frontend/src/graphql/route/Action/getRoutes.action.tsx
@@ -1,7 +1,7 @@
-import { gql, DocumentNode } from "@apollo/client";
+import { gql } from "@apollo/client";
 
 export const GET_ROUTES_QUERY = gql`
-  query GetRoutes($query: String, $limit: Int, $offset: Int) {
+  query GetRoutes($query: String, $limit: Float, $offset: Float) {
     routes(query: $query, limit: $limit, offset: $offset) {
       id
       name
